+++ CHANGELOG.rst
@@ -26,7 +26,7 @@ New features and notable changes:
   - Always sort at the end by filename if line or branch coverage is identical for a file.
   - Add :option:`--sort-casefold` to ignore the case and fold special characters.
   - Add :option:`--sort-branches` to sort by branches instead of lines, this is the default if :option:`--branches` is used.
-  - Add :option:`--sort-decreasing` to start with the files with lower coverage values.
+  - Add :option:`--sort-reverse` to reverse the sort order.
 
 Bug fixes and small improvements: